refactor(toolset): move persist_and_envelope onto ToolInvocations service
The dispatcher's `persist_and_envelope` and `envelope_text` helpers
had `&ToolInvocations` as their first arg — that's the load-bearing
shape of a method on the service, not a free function. Move them onto
`impl ToolInvocations` (and into the rendering module's neighbourhood
in `tool_invocations/mod.rs`) so the dispatcher's call site is the
expected `invocations.persist_and_envelope(subject, ...)`.

Net effect on `core/src/toolset/mod.rs`: ~150 LoC of helper code
moves out of the dispatcher file into the service it was always
operating on; the dispatcher is now purely about classifier dispatch
+ audit recording. Imports trimmed accordingly (`Content`,
`ToolInvocationId` no longer needed at the dispatcher level).
